Factory partially integrated, and playerWeapon added

+++ DynamicComponent.cpp
@@ -5,7 +5,7 @@
 #include "DynamicComponent.h"
 
 
-DynamicComponent:: DynamicComponent(int s, orientation dir): speed(s), isDestroyed(false), direction(dir){}
+DynamicComponent:: DynamicComponent(float s, orientation dir): speed(s), isDestroyed(false), direction(dir){}
 
 DynamicComponent::orientation DynamicComponent::getDirection() const {
     return direction;
